Move (de)normalize page path utils to a file `page-path-utils.ts`
