+++ packages/next/export/worker.ts
@@ -13,7 +13,7 @@ import { loadComponents } from '../server/load-components'
 import { isDynamicRoute } from '../shared/lib/router/utils/is-dynamic'
 import { getRouteMatcher } from '../shared/lib/router/utils/route-matcher'
 import { getRouteRegex } from '../shared/lib/router/utils/route-regex'
-import { normalizePagePath } from '../server/normalize-page-path'
+import { normalizePagePath } from '../server/page-path-utils'
 import { SERVER_PROPS_EXPORT_ERROR } from '../lib/constants'
 import '../server/node-polyfill-fetch'
 import { requireFontManifest } from '../server/require'
