+++ packages/next/server/dev/on-demand-entry-handler.ts
@@ -5,7 +5,7 @@ import { EventEmitter } from 'events'
 import { findPageFile } from '../lib/find-page-file'
 import { getPageRuntime, runDependingOnPageType } from '../../build/entries'
 import { join, posix } from 'path'
-import { normalizePagePath, normalizePathSep } from '../normalize-page-path'
+import { normalizePagePath, normalizePathSep } from '../page-path-utils'
 import { pageNotFoundError } from '../require'
 import { reportTrigger } from '../../build/output'
 import getRouteFromEntrypoint from '../get-route-from-entrypoint'
