+++ packages/next/server/next-server.ts
@@ -55,7 +55,7 @@ import BaseServer, {
   stringifyQuery,
 } from './base-server'
 import { getMiddlewareInfo, getPagePath, requireFontManifest } from './require'
-import { normalizePagePath } from './normalize-page-path'
+import { normalizePagePath } from './page-path-utils'
 import { loadComponents } from './load-components'
 import isError, { getProperError } from '../lib/is-error'
 import { FontManifest } from './font-utils'
